Re-organize unit tests; Move/add markdown templates for contribution
