+++ .github/ISSUE_TEMPLATE.md
@@ -1,0 +1,30 @@
+### Setup
+
+- MW version:
+- DB (MySQL etc.):
+- PHP version:
+- SMW version:
+- SRF version:
+- Browsers and versions (if applicable):
+
+### Issue
+
+...context and general description...
+
+**Steps to reproduce**
+
+(recommendation is to use the [sandbox](https://sandbox.semantic-mediawiki.org))
+
+1. ... 
+2. ...
+3. ...
+
+**Expected result:** ...
+
+**Observed result:** ...
+
+**Stacktrace** 
+
+```
+...
+```
