+++ .github/PULL_REQUEST_TEMPLATE.md
@@ -1,0 +1,10 @@
+This PR is made in reference to: #
+
+This PR addresses or contains:
+- ...
+- ...
+- ...
+
+This PR includes:
+- [ ] Tests (unit/integration)
+- [ ] CI build passed
