--- ISSUE_TEMPLATE.md
@@ -1,17 +1,0 @@
-### Setup
-
-- MW version:
-- DB (MySQL etc.):
-- PHP version:
-- SMW version:
-- SRF version:
-- Browsers and versions (if applicable):
-
-### Issue
-
-Detailed description of the issue and a [stack trace](https://www.semantic-mediawiki.org/wiki/Help:Identifying_bugs) if applicable:
-
-```
-```
-
-Steps to reproduce the observation (recommendation is to use the [sandbox](https://sandbox.semantic-mediawiki.org)):
