+++ tests/phpunit/Unit/Formats/ExcelTest.php
@@ -1,13 +1,12 @@
 <?php
 
-namespace SRF\Test;
+namespace SRF\Tests\Unit\Formats;
 
 use SMW\Test\QueryPrinterRegistryTestCase;
 
 /**
- * Tests for the SRF\DataTables class.
+ * Tests for the SRF\Excel class.
  *
- * @file
  * @since 1.9
  *
  * @ingroup SemanticResultFormats
@@ -17,31 +16,30 @@ use SMW\Test\QueryPrinterRegistryTestCase;
  * @group SMWExtension
  * @group ResultPrinters
  *
- * @licence GNU GPL v2+
- * @author mwjames
+ * @author Kim Eik
  */
-class DataTablesTest extends QueryPrinterRegistryTestCase {
+class ExcelTest extends QueryPrinterRegistryTestCase {
 
 	/**
 	 * @see QueryPrinterRegistryTestCase::getFormats
 	 *
-	 * @since 1.9
+	 * @since 1.8
 	 *
 	 * @return array
 	 */
 	public function getFormats() {
-		return [ 'datatables' ];
+		return [ 'excel' ];
 	}
 
 	/**
 	 * @see QueryPrinterRegistryTestCase::getClass
 	 *
-	 * @since 1.9
+	 * @since 1.8
 	 *
 	 * @return string
 	 */
 	public function getClass() {
-		return 'SRF\DataTables';
+		return 'SRF\SRFExcel';
 	}
 
 }
