+++ tests/phpunit/Unit/Formats/GalleryTest.php
@@ -1,11 +1,11 @@
 <?php
 
-namespace SRF\Test;
+namespace SRF\Tests\Unit\Formats;
 
 use SMW\Test\QueryPrinterRegistryTestCase;
 
 /**
- *  Tests for the SRF\Math class.
+ * Tests for the SRF\Gallery class.
  *
  * @file
  * @since 1.8
@@ -20,7 +20,7 @@ use SMW\Test\QueryPrinterRegistryTestCase;
  * @licence GNU GPL v2+
  * @author Jeroen De Dauw < jeroendedauw@gmail.com >
  */
-class MathTest extends QueryPrinterRegistryTestCase {
+class GalleryTest extends QueryPrinterRegistryTestCase {
 
 	/**
 	 * @see QueryPrinterRegistryTestCase::getFormats
@@ -30,7 +30,7 @@ class MathTest extends QueryPrinterRegistryTestCase {
 	 * @return array
 	 */
 	public function getFormats() {
-		return [ 'max', 'min', 'sum', 'product', 'average', 'median' ];
+		return [ 'gallery' ];
 	}
 
 	/**
@@ -41,7 +41,7 @@ class MathTest extends QueryPrinterRegistryTestCase {
 	 * @return string
 	 */
 	public function getClass() {
-		return '\SRFMath';
+		return 'SRF\Gallery';
 	}
 
 }
