+++ tests/phpunit/Unit/Formats/MathTest.php
@@ -1,11 +1,11 @@
 <?php
 
-namespace SRF\Test;
+namespace SRF\Tests\Unit\Formats;
 
 use SMW\Test\QueryPrinterRegistryTestCase;
 
 /**
- *  Tests for the SRF\ListWidget class.
+ *  Tests for the SRF\Math class.
  *
  * @file
  * @since 1.8
@@ -18,9 +18,9 @@ use SMW\Test\QueryPrinterRegistryTestCase;
  * @group ResultPrinters
  *
  * @licence GNU GPL v2+
- * @author mwjames
+ * @author Jeroen De Dauw < jeroendedauw@gmail.com >
  */
-class ListWidgetTest extends QueryPrinterRegistryTestCase {
+class MathTest extends QueryPrinterRegistryTestCase {
 
 	/**
 	 * @see QueryPrinterRegistryTestCase::getFormats
@@ -30,7 +30,7 @@ class ListWidgetTest extends QueryPrinterRegistryTestCase {
 	 * @return array
 	 */
 	public function getFormats() {
-		return [ 'listwidget' ];
+		return [ 'max', 'min', 'sum', 'product', 'average', 'median' ];
 	}
 
 	/**
@@ -41,7 +41,7 @@ class ListWidgetTest extends QueryPrinterRegistryTestCase {
 	 * @return string
 	 */
 	public function getClass() {
-		return '\SRFListWidget';
+		return '\SRFMath';
 	}
 
 }
