+++ tests/phpunit/Unit/Formats/SparklineTest.php
@@ -1,11 +1,11 @@
 <?php
 
-namespace SRF\Test;
+namespace SRF\Tests\Unit\Formats;
 
 use SMW\Test\QueryPrinterRegistryTestCase;
 
 /**
- * Tests for the SRF\Incoming class.
+ * Tests for the SRF\Sparkline class.
  *
  * @file
  * @since 1.8
@@ -20,7 +20,7 @@ use SMW\Test\QueryPrinterRegistryTestCase;
  * @licence GNU GPL v2+
  * @author mwjames
  */
-class IncomingTest extends QueryPrinterRegistryTestCase {
+class SparklineTest extends QueryPrinterRegistryTestCase {
 
 	/**
 	 * @see QueryPrinterRegistryTestCase::getFormats
@@ -30,7 +30,7 @@ class IncomingTest extends QueryPrinterRegistryTestCase {
 	 * @return array
 	 */
 	public function getFormats() {
-		return [ 'incoming' ];
+		return [ 'sparkline' ];
 	}
 
 	/**
@@ -41,7 +41,7 @@ class IncomingTest extends QueryPrinterRegistryTestCase {
 	 * @return string
 	 */
 	public function getClass() {
-		return '\SRFIncoming';
+		return '\SRFSparkline';
 	}
 
 }
